test: separate RPC command descriptor contracts
